Introduces specs for the Article model to ensure that the default Category "Uncategorised" is assigned to a new Article if none are specified.
 modified: spec/models/article_spec.rb
 modified: spec/support/shared_examples_grouping.rb - turns off the :check_default_category
            callback for the .with_article and nav link helpers.

The culprit is this code (and others)

  let!(:second_group) do
      create(factory, display_name: 'Second Group',
        articles: pub_articles)
  end

When the grouping is created with specific articles, the articles are saved first (even when
using the :build strategy).  Because they do not yet have a Category (it is unsaved) they are
assigned "Uncategorised" in addition to the original Category (in this case :second_group).

Skipping the callback allows the Articles to be created without any Category, which is then
assigned when the Category is saved.

* This should be avoidable using FactoryGirl callbacks to create Articles on a parent Category
so that the default "Uncategorised" Category is not assigned.